client always control only one fleet

+++ common/src/res_times.rs
@@ -1,5 +1,5 @@
 pub struct TimeRes {
-    pub tick: u64,
+    pub tick: u32,
 }
 impl TimeRes {
     pub fn new() -> Self {
